backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
libsql-experimental
firebase-admin==6.4.0
pydantic==2.6.3
python-jose[cryptography]==3.3.0
python-multipart==0.0.9 